Score now displays on game board
Unfortunately it literally displays the score on the game board. If placed outside of it, all the squares become off center. This seems to be similar to the off-center phenomena when using larger grid sizes. If I can fix that I might be able to fix this and display the squares outside the game board